requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
django-environ==0.12.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0
django-debug-toolbar==5.0.0
django-auditlog==3.2.1
